requirements.txt
openai-whisper
nemo_toolkit[asr]
evaluate
matplotlib
pandas
gTTS
ffmpeg-python
soundfile
numpy
